Separate consolidation modal and rename vars
